[ADD] Register Component IT
